fix: speed up loading the page by decoding data-files as they arrives
Before this commit, the data-files were first loaded into memory, then decoded.
But as the typeDogma is huge, this meant a lot of time the browser was sitting
idle while this file was fetched. By already starting the decoding when the
file is being read, the results are available almost as soon as the loading
finished.

The total CPU time remains the same, ofc. It just appears quicker as the
browser doesn't sit idle.